Migrate new test
Edited in https://github.com/guardian/frontend/commit/8fd9415581f3712ca332c9f7e3a1ef1012de7289
